requirements.txt
python-fasthtml>=0.10.0
faststrap
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
jinja2>=3.1.0
itsdangerous>=2.0.0
uuid>=1.30
